# backend/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WebTui Chat Backend API
  version: 0.9.0
  description: Contract nền cho backend WebTui Chat. Tài liệu API được cập nhật theo từng phase backend.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/v1/workspaces/{workspace_id}:
    get:
      summary: Xem chi tiết workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Chi tiết workspace
    patch:
      summary: Cập nhật workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Workspace đã cập nhật
    delete:
      summary: Xóa mềm workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "204":
          description: Đã xóa mềm
...
  /api/v1/workspaces/{workspace_id}/bots:
    get:
      summary: Danh sách bot trong workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Danh sách bot
    post:
      summary: Tạo bot trong workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBotRequest"
      responses:
        "201":
          description: Bot đã được tạo
  /api/v1/workspaces/{workspace_id}/bots/{bot_id}/ai-config:
    get:
      summary: Lấy cấu hình AI của bot trong workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/BotID"
      responses:
        "200":
          description: Cấu hình AI của bot
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/BotAIConfig"
    put:
      summary: Tạo hoặc cập nhật cấu hình AI của bot
      description: Mobile/web chỉ gửi secret_ref, không gửi secret/token provider thô.
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/BotID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotAIConfigRequest"
      responses:
        "200":
          description: Cấu hình AI đã được lưu
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/BotAIConfig"
  /api/v1/workspaces/{workspace_id}/bots/{bot_id}/flows:
    get:
      summary: Danh sách flow của bot trong workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/BotID"
      responses:
        "200":
          description: Danh sách flow
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          flows:
                            type: array
                            items:
                              $ref: "#/components/schemas/BotFlow"
    post:
      summary: Tạo draft flow cho bot
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/BotID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotFlowRequest"
      responses:
        "201":
          description: Flow draft đã được tạo
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/BotFlow"
  /api/v1/workspaces/{workspace_id}/bots/{bot_id}/flows/{flow_id}:
    patch:
      summary: Cập nhật draft flow của bot
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/BotID"
        - name: flow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotFlowRequest"
      responses:
        "200":
          description: Flow đã được cập nhật
  /api/v1/workspaces/{workspace_id}/bots/{bot_id}/flows/{flow_id}/publish:
    post:
      summary: Publish flow cho bot
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/BotID"
        - name: flow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Flow đã được publish
  /api/v1/workspaces/{workspace_id}/bots/{bot_id}/flows/{flow_id}/test:
    post:
      summary: Tạo run kiểm thử flow
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/BotID"
        - name: flow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotFlowTestRequest"
      responses:
        "201":
          description: Run kiểm thử flow đã được tạo
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/BotFlowRun"
  /api/v1/workspaces/{workspace_id}/bots/{bot_id}/installations:
    get:
      summary: Danh sách nơi bot được cài đặt
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/BotID"
      responses:
        "200":
          description: Danh sách cài đặt bot
    post:
      summary: Cài bot vào workspace hoặc một channel
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/BotID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallBotRequest"
      responses:
        "201":
          description: Bot đã được cài đặt
  /api/v1/workspaces/{workspace_id}/bots/{bot_id}/messages:
    post:
      summary: Gửi bot message vào channel
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/BotID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotMessageRequest"
      responses:
        "201":
          description: Bot message đã được gửi
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    WorkspaceID:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    BotID:
      name: bot_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          description: Dữ liệu trả về
        error:
          $ref: "#/components/schemas/ApiError"
        meta:
          description: Metadata phân trang hoặc thông tin bổ sung
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    CreateBotRequest:
      type: object
      required: [slug, name]
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        avatar_url:
          type: string
        settings:
          type: object
          additionalProperties: true
    InstallBotRequest:
      type: object
      properties:
        channel_id:
          type: string
          format: uuid
          description: Không gửi để cài bot ở phạm vi workspace.
        config:
          type: object
          additionalProperties: true
    BotMessageRequest:
      type: object
      required: [channel_id, body]
      properties:
        channel_id:
          type: string
          format: uuid
        body:
          type: string
        metadata:
          type: object
          additionalProperties: true
    BotAIConfig:
      type: object
      properties:
        workspace_id:
          type: string
          format: uuid
        bot_id:
          type: string
          format: uuid
        provider:
          type: string
          example: openai
        model:
          type: string
          example: gpt-4.1-mini
        secret_ref:
          type: string
          description: Tham chiếu secret trong backend/vault, không phải secret thô.
        settings:
          type: object
          additionalProperties: true
        created_by:
          type: string
          format: uuid
        updated_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BotAIConfigRequest:
      type: object
      required: [provider, model]
      properties:
        provider:
          type: string
          example: openai
        model:
          type: string
          example: gpt-4.1-mini
        secret_ref:
          type: string
          description: Tham chiếu secret đã lưu ở backend/vault.
        settings:
          type: object
          additionalProperties: true
    BotFlow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        bot_id:
          type: string
          format: uuid
        version:
          type: integer
        status:
          type: string
          enum: [draft, published, archived]
        name:
          type: string
        prompt:
          type: string
        trigger_config:
          type: object
          additionalProperties: true
        tool_config:
          type: object
          additionalProperties: true
        knowledge_config:
          type: object
          additionalProperties: true
        created_by:
          type: string
          format: uuid
        updated_by:
          type: string
          format: uuid
        published_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BotFlowRequest:
      type: object
      required: [name, prompt]
      properties:
        name:
          type: string
          maxLength: 120
        prompt:
          type: string
          maxLength: 12000
        trigger_config:
          type: object
          additionalProperties: true
        tool_config:
          type: object
          additionalProperties: true
        knowledge_config:
          type: object
          additionalProperties: true
    BotFlowTestRequest:
      type: object
      properties:
        input:
          type: object
          additionalProperties: true
    BotFlowRun:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        bot_id:
          type: string
          format: uuid
        flow_id:
          type: string
          format: uuid
        input:
          type: object
          additionalProperties: true
        transcript:
          type: object
          additionalProperties: true
        status:
          type: string
        error:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...